Initial commit
Hard to explain without the visuals, but here i go.

Version 2020.2.1f1 of Unity.

For best viewing quality best displayed in 16:9 portait resolution, but too much or too little verticality fucks the layout up. Working on that.

As you start, it generates you an enemy. You can click anywhere on the screen, and it starts the fight. It rolls d20s, adds speed, and who gets more goes first. As of now hits are all of specific value and because of that hero always wins in two shots. After that it rolls a random timer, which says how long will the wait be until the next fight will be found. When the enemy is found, you can again click on the screen to start the fight.

Healing button on the bottom heals a lot and has a 60 second cooldown.

There are upgrades on the right bar, but as of now buttons are not connected to the actual upgrade systm.

After each won fight you get coins to the amount of the IncomeIncrease amount(as of now- 1). You can also click the right button on the bottom bar, to give yourself a lot of coins, used for testing only as of now. Nowhere to spend coins, because you know, upgrades don't work yet.